+++ src/Rebus.Tests/TestStuff.cs
@@ -11,12 +11,77 @@ using Rebus.Tests.Transports.Rabbit;
 using Rebus.Transports.Msmq;
 using Shouldly;
 using System.Linq;
+using Castle.MicroKernel.Registration;
+using Castle.Windsor;
 
 namespace Rebus.Tests
 {
     [TestFixture]
     public class TestStuff
     {
+        [Test]
+        public void TestCastleWindsorFactoryMethodInjectionAndDisposal()
+        {
+            var container = new WindsorContainer()
+                .Register(
+                    Component.For<HasDependencyOnMessageContext>().LifestyleTransient(),
+
+                    Component.For<IMessageContext>()
+                        .UsingFactoryMethod(k => new FakeMessageContext())
+                        .LifestyleTransient()
+                );
+
+            container.Release(container.Resolve<HasDependencyOnMessageContext>());
+        }
+
+        class FakeMessageContext : IMessageContext
+        {
+            public void Dispose()
+            {
+                Console.WriteLine("Disposed");
+            }
+
+            public string ReturnAddress
+            {
+                get { throw new NotImplementedException(); }
+            }
+
+            public string RebusTransportMessageId
+            {
+                get { throw new NotImplementedException(); }
+            }
+
+            public IDictionary<string, object> Items
+            {
+                get { throw new NotImplementedException(); }
+            }
+
+            public void Abort()
+            {
+                throw new NotImplementedException();
+            }
+
+            public event Action Disposed;
+
+            public object CurrentMessage
+            {
+                get { throw new NotImplementedException(); }
+            }
+
+            public IDictionary<string, object> Headers
+            {
+                get { throw new NotImplementedException(); }
+            }
+        }
+
+        class HasDependencyOnMessageContext
+        {
+            public HasDependencyOnMessageContext(IMessageContext context)
+            {
+                
+            }
+        }
+
         [Test, Ignore("Only run this bad boy if you know what you're doing :)")]
         public void DeleteMsmqMessageQueuesOnTheLocalSystem()
         {