+++ CHANGELOG.md
@@ -447,6 +447,7 @@
 ## 0.75.0
 
 * Fixed merge bug that made published Rebus core assembly unusable!
+* Fixed container adapter registration bug that would make it impossible to have `IMessageContext` injected into more than one handler in the pipeline. Also ensures that the container does not dispose the message context.
 
 
 [tiipe]: https://github.com/tiipe